make module executable
